Prepare the TCK tree for JUnit 5.12
We cannot override a static field anymore and... that being said,
considering the failures I see, I think it wasn't working very well
before.

The following core PR is required for experimenting with this work:
https://github.com/quarkusio/quarkus/pull/46871

A good example of this problem is JAXRSProvidersClient0098:

- The test should fail with a NoClassDefFoundError as
  `getPropertiesTest()` in the parent class `JAXRSClient0128` is
  relying on `....core.application.TSAppConfig` being present and this
  test includes `....ext.providers.TSAppConfig` instead.
- But the test doesn't fail.
- In the Jakarta REST TCK, they actually broke the hierarchy in the
  latest version and `JAXRSProvidersClient0098` is now depending on
  `JAXRSCommonClient` instead

-> I applied the same change here.

BUT... there are a lot more failures we need to diagnose. Some of them
are quite similar so I suspect there aren't that many issues.