Prepare TCK for JUnit 5.12
Since JUnit 5.11, the rules for resolving extensions
as static fields of a class have changed.
We overcome this by not having tests whose
hierarchy attempts to create QuarkusUnitTest
in multiple places.
While doing this work we figured out that some
tests of the TCK were actually always failing
but their execution was being skipped because
of the effects of TckExtention